rustc_proc_macro

Module diagnostic

Macros

diagnostic_child_methods ๐

Structs

ChildrenExperimental

Iterator over the children diagnostics of a Diagnostic.

DiagnosticExperimental

A structure representing a diagnostic message and associated children messages.

Enums

LevelExperimental

An enum representing a diagnostic level.

Traits

MultiSpanExperimental

Trait implemented by types that can be converted into a set of Spans.